requirements.txt
flask==3.0.2
google-generativeai==0.3.2
python-dotenv==1.0.1
supabase==2.15.0
requests==2.31.0
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1 